// src/app/models/sanphamsize.ts
export class SanPhamSize{
    maSP?: string;
    tenSP?: string;
    gioiTinh?: string;
    trangThai?: string;
    kieuDang?: string;
    giaSP?: number;
    chatLieu?: string;
    mauSac?: string;
    image_URL?: string;
    soLuong38?: number;
    soLuong39?: number;
    soLuong40?: number;
    soLuong41?: number;
    soLuong42?: number;
    soLuong43?: number;
}